Bug for sign up maybe because of integer type

+++ app/models/user.py
@@ -13,7 +13,7 @@ class User(db.Model, UserMixin):
     first_name = db.Column(db.String(),nullable=False,unique=False)
     last_name = db.Column(db.String(),nullable=False,unique=False)
     balance = db.Column(db.Integer(),nullable=False,default=0)
-    phone_number = db.Column(db.Integer(),nullable=False,unique=True)
+    phone_number = db.Column(db.Numeric(10),nullable=False,unique=True)
     username = db.Column(db.String(40), nullable=False, unique=True)
     email = db.Column(db.String(255), nullable=False, unique=True)
     hashed_password = db.Column(db.String(255), nullable=False)
